docs: Add development checklist
