+++ aiblog-client/src/pages/MainPage.jsx
@@ -2,7 +2,7 @@ import { Box, Container } from '@mui/material';
 import { RepoSetup } from '@/components/RepoSetup';
 import { LeftPanel } from '@/components/LeftPanel';
 import { RightPanel } from '@/components/RightPanel';
-import { GenerateButton } from '@/components/GenerateButton'; // Import here
+import { GenerateButton } from '@/components/GenerateButton';
 import { MainPageProvider, useMainPageContext } from '@/contexts/MainPageContext';
 
 function MainPageContent() {
