+++ specification_week2.md
@@ -1,0 +1,16 @@
+학습 목표
+* 다양한 외부 API를 연동해서 서비스를 구현할 수 있다.
+* LLM을 연동한 AI 서비스를 만들 수 있다.
+* AI를 활용해서 애플리케이션을 설계하고 구현할 수 있다.
+기능요구사항
+2주차 미션 : LLM 연동
+* GitHub 최근 작업(커밋/PR) 별로 '블로그생성' 버튼을 누르면 LLM연동을 통해서 블로그가 생성된다.
+* 생성된 블로그 내용이 화면에 노출된다.
+* UI 표현 방식은 자유롭게 구성한다.
+프로그래밍 요구사항
+LLM 연동
+* 블로그 글을 자동으로 작성하기 위해서 LLM서비스를 연동한다.
+* LLM 모델 선택은 자유다.(Gemini등)
+* LLM API 연동 작업은 서버에서 진행한다.
+* token및 기타 인증정보를 유출하지 않도록 조심한다.
+* 블로그 결과는 React를 통해서 자유롭게 렌더링해서 노출한다.
